@@ -34,34 +34,32 @@
     "@google-cloud/storage": "^7.7.0",
     "@hookform/resolvers": "^3.3.4",
     "@next-auth/prisma-adapter": "^1.0.7",
-    "@prisma/client": "5.10.2",
+    "@prisma/client": "5.16.1",
     "@radix-ui/colors": "^3.0.0",
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-popover": "^1.1.1",
     "@radix-ui/react-select": "^2.0.0",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-tooltip": "^1.0.7",
     "@sentry/nextjs": "^7.102.1",
     "@slack/webhook": "^7.0.1",
-    "@stripe/stripe-js": "^2.4.0",
-    "@tanstack/react-query": "^5.20.1",
-    "@uiw/react-color-sketch": "^2.0.8",
+    "@stripe/stripe-js": "^4.1.0",
     "@uiw/react-markdown-preview": "^5.0.6",
     "@uiw/react-md-editor": "^4.0.3",
-    "axios": "^1.5.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "formidable": "^3.5.1",
     "jotai": "^2.6.4",
-    "lucide-react": "^0.344.0",
-    "next": "14.1.0",
-    "next-auth": "^4.24.5",
-    "next-connect": "^0.13.0",
+    "lucide-react": "^0.438.0",
+    "next": "14.2.4",
+    "next-auth": "^4.24.7",
     "next-remove-imports": "^1.0.12",
-    "next-themes": "^0.2.1",
+    "next-safe-action": "^7.0.2",
+    "next-themes": "^0.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-dropzone": "^14.2.3",
@@ -70,39 +68,48 @@
     "react-paginate": "^8.2.0",
     "react-papaparse": "^4.4.0",
     "rehype-sanitize": "^6.0.0",
-    "resend": "^3.2.0",
+    "resend": "^4.0.0",
     "slugify": "^1.6.6",
     "sonner": "^1.4.0",
-    "stripe": "^14.16.0",
+    "stripe": "^16.1.0",
     "tailwind-merge": "^2.2.0",
     "tailwindcss-animate": "^1.0.6",
-    "uuid": "^9.0.1",
+    "use-debounce": "^10.0.0",
     "vaul": "^0.9.0",
-    "wcag-contrast": "^3.0.0",
     "zod": "^3.22.4"
   },
   "devDependencies": {
     "@changesets/cli": "^2.27.1",
-    "@playwright/test": "^1.41.1",
-    "@swc-jotai/react-refresh": "^0.1.0",
+    "@eslint/compat": "^1.1.1",
+    "@eslint/eslintrc": "^3.1.0",
+    "@eslint/js": "^9.9.1",
+    "@next/eslint-plugin-next": "^14.2.7",
+    "@playwright/test": "^1.45.0",
+    "@swc-jotai/react-refresh": "^0.2.0",
     "@types/formidable": "^3.4.5",
-    "@types/node": "^20.5.6",
+    "@types/node": "^22.5.4",
     "@types/react": "^18.0.27",
     "autoprefixer": "^10.4.16",
     "dotenv": "^16.4.5",
-    "eslint": "8.57.0",
-    "eslint-config-next": "14.1.0",
+    "eslint": "9.9.1",
+    "eslint-config-next": "^14.2.7",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-unused-imports": "^3.0.0",
+    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-plugin-import-x": "^4.2.0",
+    "eslint-plugin-playwright": "^1.6.2",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.35.2",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-unused-imports": "^4.1.3",
     "husky": "^9.0.11",
     "lint-staged": "^15.0.2",
     "postcss": "^8.4.28",
     "prettier": "^3.2.5",
-    "prettier-plugin-tailwindcss": "^0.5.11",
-    "prisma": "5.10.2",
+    "prettier-plugin-tailwindcss": "^0.6.6",
+    "prisma": "5.16.1",
     "tailwindcss": "^3.4.1",
     "ts-node": "^10.9.1",
-    "typescript": "^5.3.3"
+    "typescript": "^5.5.4",
+    "typescript-eslint": "^8.4.0"
   }
 }
