@@ -1,9 +1,10 @@
-import { Dispatch, MouseEvent, SetStateAction, useState } from 'react';
+'use client';
 
-import { Tag } from '@prisma/client';
+import type { MouseEvent } from 'react';
+
 import { SearchIcon, TagIcon } from 'lucide-react';
-import { useRouter } from 'next/router';
-import { useForm } from 'react-hook-form';
+import { usePathname, useRouter, useSearchParams } from 'next/navigation';
+import { useDebouncedCallback } from 'use-debounce';
 
 import {
   DropdownMenu,
@@ -14,53 +15,44 @@ import {
   Label,
 } from '@/components';
 
+import type { Tag } from '@prisma/client';
+
 type Props = {
   tags: Tag[];
-  setSearchQuery: Dispatch<SetStateAction<string>>;
-  setSearchTag: Dispatch<SetStateAction<string>>;
-  setPage: Dispatch<SetStateAction<number>>;
 };
 
-export const Search = ({
-  tags,
-  setSearchQuery,
-  setSearchTag,
-  setPage,
-}: Props) => {
-  const router = useRouter();
-  const { handleSubmit, register } = useForm();
-  const [tagActive, setTagActive] = useState<string | null>(null);
+export const Search = ({ tags }: Props) => {
+  const searchParams = useSearchParams();
+  const pathname = usePathname();
+  const { replace } = useRouter();
+  const params = new URLSearchParams(searchParams);
+  const currentTag = params.get('tag') || '';
 
-  const onSearch = (values) => {
-    const { search } = values;
-    setSearchQuery(search);
-    setPage(0);
+  const handleSearch = useDebouncedCallback((search) => {
+    params.delete('page');
     if (search) {
-      const encodedSearch = encodeURI(search);
-      router.push(`?search=${encodedSearch}`);
+      params.set('query', search);
     } else {
-      router.push('');
+      params.delete('query');
     }
-  };
+    replace(`${pathname}?${params.toString()}`);
+  }, 300);
 
   const handleTagSearch = (label: string) => {
-    setTagActive(label);
-    setSearchTag(label);
-    setPage(0);
+    params.delete('page');
+    params.set('tag', label);
+    replace(`${pathname}?${params.toString()}`);
   };
 
   const handleResetTag = () => {
-    setTagActive(null);
-    setSearchTag(null);
-    setPage(0);
+    params.delete('page');
+    params.delete('tag');
+    replace(`${pathname}?${params.toString()}`);
   };
 
   return (
     <section className="mx-auto flex w-11/12 items-end justify-center gap-8 md:w-3/4">
-      <form
-        onSubmit={handleSubmit(onSearch)}
-        className="group/search flex w-full flex-col gap-1 [&_svg]:focus-within:text-tealA-8"
-      >
+      <div className="group/search flex w-full flex-col gap-1 [&_svg]:focus-within:text-tealA-8">
         <Label
           htmlFor="search"
           className="lowercase"
@@ -69,15 +61,16 @@ export const Search = ({
           Search
         </Label>
         <Input
-          {...register('search')}
           withIcon
           icon={<SearchIcon />}
-          type="text"
+          type="search"
           id="search"
-          placeholder="Search"
+          placeholder="Question title"
           className="py-2"
+          defaultValue={searchParams.get('query')?.toString()}
+          onChange={(e) => handleSearch(e.target.value)}
         />
-      </form>
+      </div>
       {tags.length > 0 && (
         <DropdownMenu>
           <DropdownMenuTrigger
@@ -90,22 +83,25 @@ export const Search = ({
             {tags.map((tag) => (
               <DropdownMenuItem
                 className={
-                  tagActive === tag.label
+                  currentTag === tag.label
                     ? 'bg-gray-12 text-white hover:bg-gray-12 dark:bg-gray-1 dark:hover:bg-gray-1'
                     : 'bg-gray-1 dark:bg-gray-12'
                 }
                 key={tag.id}
               >
                 <button
-                  className="h-full w-full text-start"
+                  className="size-full text-start"
+                  type="button"
                   onClick={(e: MouseEvent<HTMLButtonElement>) =>
                     handleTagSearch(e.currentTarget.innerHTML)
                   }
                 >
                   {tag.label}
                 </button>
-                {tagActive === tag.label && (
-                  <button onClick={handleResetTag}>x</button>
+                {currentTag === tag.label && (
+                  <button onClick={handleResetTag} type="button">
+                    x
+                  </button>
                 )}
               </DropdownMenuItem>
             ))}
