@@ -8,34 +8,33 @@ const nextConfig = removeImports({
   images: {
     domains: [
       'lh3.googleusercontent.com',
-      'api.dicebear.com',
       'storage.googleapis.com',
     ],
   },
-  experimental: {
-    swcPlugins: [['@swc-jotai/react-refresh', {}]],
-  },
+  // experimental: {
+  //   swcPlugins: [['@swc-jotai/react-refresh', {}]],
+  // },
   webpack: (config) => {
     config.experiments = { ...config.experiments, topLevelAwait: true };
     return config;
   },
 });
 
-const sentryConfig = withSentryConfig(
-  nextConfig,
-  {
-    authToken: process.env.SENTRY_AUTH_TOKEN,
-    silent: true,
-    org: 'thibaud-brault',
-    project: 'faqmaker',
-  },
-  {
-    widenClientFileUpload: true,
-    transpileClientSDK: true,
-    tunnelRoute: '/monitoring',
-    hideSourceMaps: true,
-    disableLogger: true,
-  },
-);
+// const sentryConfig = withSentryConfig(
+//   nextConfig,
+//   {
+//     authToken: process.env.SENTRY_AUTH_TOKEN,
+//     silent: true,
+//     org: 'thibaud-brault',
+//     project: 'faqmaker',
+//   },
+//   {
+//     widenClientFileUpload: true,
+//     transpileClientSDK: true,
+//     tunnelRoute: '/monitoring',
+//     hideSourceMaps: true,
+//     disableLogger: true,
+//   },
+// );
 
-module.exports = sentryConfig;
+module.exports = nextConfig;
