@@ -1,4 +1,5 @@
-import { forwardRef, HTMLAttributes } from 'react';
+import type { HTMLAttributes } from 'react';
+import { forwardRef } from 'react';
 
 import { cva, type VariantProps } from 'class-variance-authority';
 
@@ -7,14 +8,14 @@ import { cn } from '@/utils';
 const badge = cva('text-xs font-semibold', {
   variants: {
     variant: {
-      primary: ['bg-gray-12 text-gray-1 text-center'],
-      disabled: ['bg-gray-11 text-gray-1 text-center'],
+      primary: ['bg-gray-12 text-center text-gray-1'],
+      disabled: ['bg-gray-11 text-center text-gray-1'],
     },
     rounded: {
       full: ['rounded-full'],
     },
     size: {
-      small: ['px-2 min-w-[40px]'],
+      small: ['min-w-[40px] px-2'],
     },
   },
 });
