@@ -6,15 +6,14 @@ export * from './drawer';
 export * from './dropdown';
 export * from './editor';
 export * from './email';
-export * from './error';
 export * from './field';
 export * from './input';
 export * from './label';
 export * from './loader';
 export * from './pagination';
+export * from './popover';
 export * from './select';
 export * from './stepper';
 export * from './tabs';
-export * from './theme';
 export * from './toast';
 export * from './tooltip';
