@@ -1,8 +1,7 @@
-import { AxiosResponse } from 'axios';
 import { toast } from 'sonner';
 
 export const errorToast = (error?: string) => {
-  return toast.error(error ? error : `Something went wrong`);
+  return toast.error(error || `Something went wrong`);
 };
 
 export const successToast = (text: string) => {
@@ -11,20 +10,33 @@ export const successToast = (text: string) => {
 };
 
 export const promiseToast = (
+  /* eslint-disable-next-line @typescript-eslint/no-explicit-any*/
   promise: Promise<any> | (() => Promise<any>),
   loading: string,
 ) => {
   return toast.promise(promise, {
-    loading: loading,
-    success: (data: AxiosResponse['data']) => {
+    loading,
+    success: (data) => {
       return `${data.message}`;
     },
     error: (data) => {
       if (data.response.data.error) {
         return `${data.response.data.error.message}`;
-      } else {
-        return 'Something went wrong';
       }
+      return 'Something went wrong';
     },
   });
+};
+
+export const resultToast = (
+  error: string | undefined,
+  success: string | undefined,
+) => {
+  if (error) {
+    return errorToast(error);
+  }
+  if (success) {
+    return successToast(success);
+  }
+  return null;
 };
