@@ -1,29 +1,20 @@
-import { Question } from '@prisma/client';
+'use client';
+
 import Link from 'next/link';
 
-import { Loader } from '@/components';
-import { dateOptions } from '@/utils';
+import { Routes, dateOptions } from '@/utils';
 
-type QuestionWithNodeId = Question & {
-  node: {
-    id: string;
-  };
-};
+import type { QuestionWithNodeId } from '@/types';
 
 type Props = {
   questions?: QuestionWithNodeId[];
-  isPending: boolean;
 };
 
-export const UserQuestions = ({ questions, isPending }: Props) => {
-  if (isPending) {
-    return <Loader size="items" />;
-  }
-
+export const UserQuestions = ({ questions }: Props) => {
   return (
     <>
       <h2
-        className="text-center font-serif text-3xl font-semibold lowercase md:text-4xl"
+        className="mb-4 text-center font-serif text-3xl font-semibold lowercase md:text-4xl"
         style={{ fontVariant: 'small-caps' }}
       >
         Questions
@@ -32,15 +23,12 @@ export const UserQuestions = ({ questions, isPending }: Props) => {
         <ul className="flex list-none flex-col gap-2">
           {questions?.map((question) => (
             <li
-              className="flex items-center justify-between rounded-md px-3 py-2 shadow-sm"
+              className="flex items-center justify-between rounded-md px-3 py-2 shadow-sm hover:shadow-teal-6"
               key={question.id}
             >
               <h3 className="text-xl font-semibold hover:underline md:text-2xl">
                 <Link
-                  href={{
-                    pathname: '/question/[id]',
-                    query: { id: question.node.id },
-                  }}
+                  href={`${Routes.SITE.QUESTION.INDEX}/${question.node.id}`}
                 >
                   {question.text}
                 </Link>
