@@ -1,45 +1,66 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
-import { Tenant } from '@prisma/client';
 import { Upload } from 'lucide-react';
 import Image from 'next/image';
 import Dropzone from 'react-dropzone';
-import { Controller, SubmitHandler, useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { Controller, useForm } from 'react-hook-form';
 
-import { Button } from '@/components';
-import { useUpsertFiles } from '@/hooks';
-import { filesClientSchema } from '@/lib';
+import { getSignedLogo, updateLogo } from '@/actions';
+import { Button, resultToast } from '@/components';
+import { filesSchema } from '@/lib/validations';
+
+import type { Tenant } from '@prisma/client';
+import type { SubmitHandler } from 'react-hook-form';
+import type { z } from 'zod';
 
 type Props = {
   tenant: Tenant;
 };
 
-type Schema = z.infer<typeof filesClientSchema>;
+type Schema = z.infer<typeof filesSchema>;
 
 export const Files = ({ tenant }: Props) => {
   const [disabled, setDisabled] = useState<boolean>(true);
   const [previewImage, setPreviewImage] = useState<string>('');
-  const [file, setFile] = useState<File>();
+  const [file, setFile] = useState<File | null>(null);
 
   const {
     handleSubmit,
     control,
     reset,
-    formState: { isSubmitting, isDirty, isValid, errors },
+    formState: { isSubmitting, isDirty, isValid },
   } = useForm<Schema>({
-    resolver: zodResolver(filesClientSchema),
+    resolver: zodResolver(filesSchema),
     mode: 'onBlur',
     defaultValues: {
-      logo: null,
+      logo: undefined,
     },
   });
 
-  const { mutate } = useUpsertFiles(tenant.id);
-
-  const onSubmit: SubmitHandler<Schema> = (values) => {
-    mutate(values);
+  const onSubmit: SubmitHandler<Schema> = async (data) => {
+    const formData = new FormData();
+    const randomId = crypto.randomUUID();
+    const logo = encodeURIComponent(randomId + data.logo.name);
+    formData.append('logo', logo);
+    const { url, fields } = await getSignedLogo(formData);
+    formData.delete('logo');
+    Object.entries({ ...fields, file }).forEach(([key, value]) => {
+      formData.append(key, value as string | Blob);
+    });
+    await fetch(url, { method: 'POST', body: formData });
+    formData.forEach((_value, key) => {
+      formData.delete(key);
+    });
+    const logoUrl = `${url}logos/${logo}`;
+    const logoData: UpdateLogo = {
+      logoUrl,
+      id: tenant.id,
+    };
+    const result = await updateLogo(logoData);
+    resultToast(result?.serverError, 'Logo updated successfully');
   };
 
   const handleReset = () => {
@@ -50,8 +71,6 @@ export const Files = ({ tenant }: Props) => {
 
   useEffect(() => {
     setDisabled(isSubmitting || !isDirty || !isValid);
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isDirty, isSubmitting, isValid]);
 
   return (
@@ -67,7 +86,7 @@ export const Files = ({ tenant }: Props) => {
         onSubmit={handleSubmit(onSubmit)}
       >
         <Controller
-          name={'logo'}
+          name="logo"
           control={control}
           render={({ field: { onChange, onBlur } }) => (
             <div className="w-full">
@@ -96,14 +115,14 @@ export const Files = ({ tenant }: Props) => {
                     {previewImage ? (
                       <Image
                         src={previewImage}
-                        alt={file.name}
-                        className="h-36 w-36 rounded-md border border-gray-7 object-cover"
+                        alt={file?.name ?? 'Logo'}
+                        className="size-36 rounded-md border border-gray-7 object-cover"
                         width={144}
                         height={144}
                       />
                     ) : (
-                      <div className="flex h-36 w-36 items-center justify-center">
-                        <Upload className="h-20 w-20" />
+                      <div className="flex size-36 items-center justify-center">
+                        <Upload className="size-20" />
                       </div>
                     )}
                     <p className="text-xl font-semibold">
