@@ -1,24 +1,25 @@
-import { Tenant } from '@prisma/client';
-
-import { Loader } from '@/components';
-
 import { Company } from './Company';
 import { Data } from './Data';
 import { Files } from './Files';
 import { Integrations } from './Integrations';
 
+import type { Integrations as IntegrationsType, Tenant } from '@prisma/client';
+
 type Props = {
-  tenantId: string;
   tenant: Tenant;
-  isPending: boolean;
+  integrations: IntegrationsType | null;
+  nodesCount: number;
+  usersCount: number;
 };
 
-export const General = ({ tenantId, tenant, isPending }: Props) => {
+export const General = ({
+  tenant,
+  integrations,
+  nodesCount,
+  usersCount,
+}: Props) => {
   const styles = 'w-full rounded-md bg-gray-3 p-4';
 
-  if (isPending) {
-    return <Loader size="page" />;
-  }
   return (
     <>
       <section className={styles}>
@@ -29,11 +30,15 @@ export const General = ({ tenantId, tenant, isPending }: Props) => {
       </section>
       {tenant.plan !== 'free' && (
         <section className={styles}>
-          <Integrations tenantId={tenantId} />
+          <Integrations tenantId={tenant.id} integrations={integrations} />
         </section>
       )}
       <section className="w-full">
-        <Data tenantId={tenantId} plan={tenant.plan} />
+        <Data
+          plan={tenant.plan}
+          nodesCount={nodesCount}
+          usersCount={usersCount}
+        />
       </section>
     </>
   );
