@@ -1,8 +1,10 @@
+'use client';
+
 import { Banknote } from 'lucide-react';
 import { useForm } from 'react-hook-form';
 
-import { Button } from '@/components';
-import { useCreateBillingPortal } from '@/hooks';
+import { Button, errorToast } from '@/components';
+import { Routes } from '@/utils';
 
 type Props = {
   tenantId: string;
@@ -10,10 +12,22 @@ type Props = {
 
 export const Billing = ({ tenantId }: Props) => {
   const { handleSubmit } = useForm();
-  const { mutate } = useCreateBillingPortal(tenantId);
 
-  const onSubmit = () => {
-    mutate();
+  const onSubmit = async () => {
+    const body = { tenantId };
+    const response = await fetch(Routes.API.BILLING, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(body),
+    });
+    const data = await response.json();
+    if (response.ok) {
+      window.location.assign(data.url);
+    } else {
+      errorToast(data.error);
+    }
   };
 
   return (
