@@ -2,9 +2,9 @@ import { useEffect, useState } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
 import { PlusCircle, Tag as TagIcon } from 'lucide-react';
-import { SubmitHandler, useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { useForm } from 'react-hook-form';
 
+import { createTag, createTagSchema } from '@/actions';
 import {
   Button,
   Dialog,
@@ -19,17 +19,87 @@ import {
   DrawerTrigger,
   Field,
   Input,
+  resultToast,
 } from '@/components';
-import { useCreateTag, useMediaQuery } from '@/hooks';
-import { createTagClientSchema } from '@/lib';
+import { useMediaQuery } from '@/hooks';
+
+import type { $Enums } from '@prisma/client';
+import type { SubmitHandler } from 'react-hook-form';
+import type { z } from 'zod';
 
 type Props = {
   tenantId: string;
+  plan: $Enums.Plan;
+  tagsCount: number;
 };
 
-type Schema = z.infer<typeof createTagClientSchema>;
+type Schema = z.infer<typeof createTagSchema>;
 
-export const CreateTag = ({ tenantId }: Props) => {
+const Form = ({ tenantId, plan, tagsCount }: Props) => {
+  const [disabled, setDisabled] = useState<boolean>(true);
+
+  const {
+    register,
+    handleSubmit,
+    watch,
+    formState: { errors, isSubmitting, isValid },
+  } = useForm<Schema>({
+    resolver: zodResolver(createTagSchema),
+    mode: 'onBlur',
+    defaultValues: {
+      label: '',
+      tagsCount,
+      plan,
+      tenantId,
+    },
+  });
+
+  const onSubmit: SubmitHandler<Schema> = async (data) => {
+    const result = await createTag(data);
+    resultToast(result?.serverError, 'Tag created successfully');
+  };
+
+  useEffect(() => {
+    setDisabled(isSubmitting || !isValid);
+  }, [isSubmitting, isValid]);
+
+  return (
+    <form
+      onSubmit={handleSubmit(onSubmit)}
+      className="flex flex-col items-center gap-2"
+    >
+      <fieldset className="mx-auto flex w-11/12 flex-col gap-1">
+        <Field
+          label="Label"
+          value="label"
+          error={errors.label?.message}
+          curLength={watch('label')?.length}
+          limit={50}
+        >
+          <Input
+            {...register('label')}
+            withIcon
+            icon={<TagIcon />}
+            type="label"
+            id="label"
+            placeholder="Tag label"
+          />
+        </Field>
+      </fieldset>
+      <Button
+        variant={disabled ? 'disabled' : 'primary'}
+        weight="semibold"
+        className="lowercase"
+        style={{ fontVariant: 'small-caps' }}
+        disabled={disabled}
+      >
+        Add
+      </Button>
+    </form>
+  );
+};
+
+export const CreateTag = ({ tenantId, plan, tagsCount }: Props) => {
   const isDesktop = useMediaQuery('(min-width: 640px)');
 
   if (isDesktop) {
@@ -53,7 +123,7 @@ export const CreateTag = ({ tenantId }: Props) => {
           <DialogHeader>
             <DialogTitle>New tag</DialogTitle>
           </DialogHeader>
-          <Form tenantId={tenantId} />
+          <Form tenantId={tenantId} plan={plan} tagsCount={tagsCount} />
         </DialogContent>
       </Dialog>
     );
@@ -80,65 +150,9 @@ export const CreateTag = ({ tenantId }: Props) => {
           <DrawerHeader>
             <DrawerTitle>New tag</DrawerTitle>
           </DrawerHeader>
-          <Form tenantId={tenantId} />
+          <Form tenantId={tenantId} plan={plan} tagsCount={tagsCount} />
         </div>
       </DrawerContent>
     </Drawer>
-  );
-};
-
-const Form = ({ tenantId }: Props) => {
-  const [disabled, setDisabled] = useState<boolean>(true);
-
-  const {
-    register,
-    handleSubmit,
-    reset,
-    formState: { errors, isSubmitting, isValid },
-  } = useForm<Schema>({
-    resolver: zodResolver(createTagClientSchema),
-    mode: 'onBlur',
-    defaultValues: {
-      label: '',
-    },
-  });
-
-  const { mutate, isError, error } = useCreateTag(tenantId, reset);
-
-  const onSubmit: SubmitHandler<Schema> = (values) => {
-    mutate(values);
-  };
-
-  useEffect(() => {
-    setDisabled(isSubmitting || !isValid);
-  }, [isSubmitting, isValid]);
-
-  return (
-    <form
-      onSubmit={handleSubmit(onSubmit)}
-      className="flex flex-col items-center gap-2"
-    >
-      <fieldset className="mx-auto flex w-11/12 flex-col gap-1">
-        <Field label="Label" value="label" error={errors.label?.message}>
-          <Input
-            {...register('label')}
-            withIcon
-            icon={<TagIcon />}
-            type="label"
-            id="label"
-            placeholder="Tag label"
-          />
-        </Field>
-      </fieldset>
-      <Button
-        variant={disabled ? 'disabled' : 'primary'}
-        weight="semibold"
-        className="lowercase"
-        style={{ fontVariant: 'small-caps' }}
-        disabled={disabled}
-      >
-        Add
-      </Button>
-    </form>
   );
 };
