@@ -1,10 +1,12 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
 import { AtSign, PlusCircle } from 'lucide-react';
-import { Controller, SubmitHandler, useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { Controller, useForm } from 'react-hook-form';
 
+import { createUser, createUserSchema } from '@/actions';
 import {
   Button,
   Dialog,
@@ -25,96 +27,44 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
+  resultToast,
 } from '@/components';
-import { useCreateUser, useMediaQuery } from '@/hooks';
-import { createUserClientSchema } from '@/lib';
+import { useMediaQuery } from '@/hooks';
+import { Limits } from '@/utils';
+
+import type { SubmitHandler } from 'react-hook-form';
+import type { z } from 'zod';
 
 type Props = {
   tenantId: string;
+  usersCount: number;
 };
 
-type Schema = z.infer<typeof createUserClientSchema>;
+type Schema = z.infer<typeof createUserSchema>;
 
-export const CreateUser = ({ tenantId }: Props) => {
-  const isDesktop = useMediaQuery('(min-width: 640px)');
-
-  if (isDesktop) {
-    return (
-      <Dialog>
-        <DialogTrigger asChild>
-          <Button
-            variant="primary"
-            size="full"
-            font="large"
-            icon="withIcon"
-            weight="semibold"
-            className="lowercase"
-            style={{ fontVariant: 'small-caps' }}
-          >
-            <PlusCircle />
-            New user
-          </Button>
-        </DialogTrigger>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>New user</DialogTitle>
-          </DialogHeader>
-          <Form tenantId={tenantId} />
-        </DialogContent>
-      </Dialog>
-    );
-  }
-
-  return (
-    <Drawer>
-      <DrawerTrigger asChild>
-        <Button
-          variant="primary"
-          size="full"
-          font="large"
-          icon="withIcon"
-          weight="semibold"
-          className="lowercase"
-          style={{ fontVariant: 'small-caps' }}
-        >
-          <PlusCircle />
-          New user
-        </Button>
-      </DrawerTrigger>
-      <DrawerContent>
-        <div className="mb-10 mt-5">
-          <DrawerHeader>
-            <DrawerTitle className="text-2xl">New user</DrawerTitle>
-          </DrawerHeader>
-          <Form tenantId={tenantId} />
-        </div>
-      </DrawerContent>
-    </Drawer>
-  );
-};
-
-const Form = ({ tenantId }: Props) => {
+const Form = ({ tenantId, usersCount }: Props) => {
   const [disabled, setDisabled] = useState<boolean>(true);
 
   const {
     register,
     handleSubmit,
     control,
-    reset,
+    watch,
     formState: { errors, isValid, isSubmitting },
   } = useForm<Schema>({
-    resolver: zodResolver(createUserClientSchema),
+    resolver: zodResolver(createUserSchema),
     mode: 'onBlur',
     defaultValues: {
       email: '',
       role: 'user',
+      tenantId,
+      usersCount,
     },
   });
 
-  const { mutate } = useCreateUser(tenantId, reset);
-
-  const onSubmit: SubmitHandler<Schema> = (values) => {
-    mutate(values);
+  const onSubmit: SubmitHandler<Schema> = async (data) => {
+    const result = await createUser(data);
+    resultToast(result?.serverError, result?.data?.message);
   };
 
   useEffect(() => {
@@ -127,11 +77,17 @@ const Form = ({ tenantId }: Props) => {
       className="flex flex-col items-center gap-4"
     >
       <fieldset className="mx-auto flex w-11/12 flex-col gap-2">
-        <Field label="Email" value="email" error={errors.email?.message}>
+        <Field
+          label="Email"
+          value="email"
+          error={errors.email?.message}
+          limit={Limits.EMAIL}
+          curLength={watch('email').length}
+        >
           <Input
             {...register('email')}
             withIcon
-            icon={<AtSign className="h-5 w-5" />}
+            icon={<AtSign className="size-5" />}
             type="email"
             id="email"
             placeholder="Email"
@@ -177,5 +133,63 @@ const Form = ({ tenantId }: Props) => {
         Add
       </Button>
     </form>
+  );
+};
+
+export const CreateUser = ({ tenantId, usersCount }: Props) => {
+  const isDesktop = useMediaQuery('(min-width: 640px)');
+
+  if (isDesktop) {
+    return (
+      <Dialog>
+        <DialogTrigger asChild>
+          <Button
+            variant="primary"
+            size="full"
+            font="large"
+            icon="withIcon"
+            weight="semibold"
+            className="lowercase"
+            style={{ fontVariant: 'small-caps' }}
+          >
+            <PlusCircle />
+            New user
+          </Button>
+        </DialogTrigger>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>New user</DialogTitle>
+          </DialogHeader>
+          <Form tenantId={tenantId} usersCount={usersCount} />
+        </DialogContent>
+      </Dialog>
+    );
+  }
+
+  return (
+    <Drawer>
+      <DrawerTrigger asChild>
+        <Button
+          variant="primary"
+          size="full"
+          font="large"
+          icon="withIcon"
+          weight="semibold"
+          className="lowercase"
+          style={{ fontVariant: 'small-caps' }}
+        >
+          <PlusCircle />
+          New user
+        </Button>
+      </DrawerTrigger>
+      <DrawerContent>
+        <div className="mb-10 mt-5">
+          <DrawerHeader>
+            <DrawerTitle className="text-2xl">New user</DrawerTitle>
+          </DrawerHeader>
+          <Form tenantId={tenantId} usersCount={usersCount} />
+        </div>
+      </DrawerContent>
+    </Drawer>
   );
 };
