@@ -1,11 +1,12 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
-import { User } from '@prisma/client';
 import { AtSign, UserIcon } from 'lucide-react';
-import { Controller, SubmitHandler, useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { Controller, useForm } from 'react-hook-form';
 
+import { updateUser, updateUserSchema } from '@/actions';
 import {
   Button,
   Dialog,
@@ -27,15 +28,122 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components';
-import { useMediaQuery, useUpdateUser } from '@/hooks';
-import { updateUserClientSchema } from '@/lib';
+import { useMediaQuery } from '@/hooks';
+
+import type { User } from '@prisma/client';
+import type { SubmitHandler } from 'react-hook-form';
+import type { z } from 'zod';
 
 type Props = {
   user: User;
   tenantId: string;
 };
 
-type Schema = z.infer<typeof updateUserClientSchema>;
+type Schema = z.infer<typeof updateUserSchema>;
+
+const Form = ({ user, tenantId }: Props) => {
+  const [disabled, setDisabled] = useState<boolean>(true);
+
+  const {
+    register,
+    handleSubmit,
+    control,
+    formState: { isSubmitting, isDirty, isValid, errors },
+  } = useForm<Schema>({
+    resolver: zodResolver(updateUserSchema),
+    mode: 'onBlur',
+    defaultValues: {
+      name: user.name ?? '',
+      email: user.email,
+      role: user.role,
+      tenantId,
+    },
+  });
+
+  const onSubmit: SubmitHandler<Schema> = async (data) => {
+    await updateUser(data);
+  };
+
+  useEffect(() => {
+    setDisabled(isSubmitting || !isDirty || !isValid);
+  }, [isDirty, isSubmitting, isValid]);
+
+  return (
+    <form
+      onSubmit={handleSubmit(onSubmit)}
+      className="flex flex-col items-center gap-4"
+    >
+      <fieldset className="mx-auto flex w-11/12 flex-col gap-2">
+        <div className="flex flex-col gap-1">
+          <Field label="Name" value="name" error={errors.name?.message}>
+            <Input
+              {...register('name')}
+              withIcon
+              defaultValue={user.name ?? ''}
+              icon={<UserIcon className="size-5" />}
+              type="name"
+              id="name"
+              placeholder="Name"
+            />
+          </Field>
+        </div>
+        <div className="flex flex-col gap-1">
+          <Field label="Email" value="email" error={errors.email?.message}>
+            <Input
+              {...register('email', { required: true })}
+              withIcon
+              defaultValue={user.email}
+              icon={<AtSign className="size-5" />}
+              type="email"
+              id="email"
+              placeholder="Email"
+            />
+          </Field>
+        </div>
+        {user.role !== 'tenant' && (
+          <div className="flex flex-col gap-1">
+            <Label
+              htmlFor="role"
+              className="lowercase"
+              style={{ fontVariant: 'small-caps' }}
+            >
+              Role
+            </Label>
+            <Controller
+              control={control}
+              name="role"
+              rules={{ required: true }}
+              render={({ field: { onChange } }) => (
+                <Select onValueChange={onChange} defaultValue={user.role}>
+                  <SelectTrigger id="role">
+                    <SelectValue placeholder="Select a role" />
+                  </SelectTrigger>
+                  <SelectContent>
+                    <SelectItem value="user" className="pl-8">
+                      User
+                    </SelectItem>
+                    <SelectItem value="admin" className="pl-8">
+                      Admin
+                    </SelectItem>
+                  </SelectContent>
+                </Select>
+              )}
+            />
+          </div>
+        )}
+      </fieldset>
+      <Button
+        variant={disabled ? 'disabled' : 'primary'}
+        weight="semibold"
+        className="lowercase"
+        style={{ fontVariant: 'small-caps' }}
+        disabled={disabled}
+      >
+        Update
+      </Button>
+    </form>
+  );
+};
 
 export const UpdateUser = ({ user, tenantId }: Props) => {
   const isDesktop = useMediaQuery('(min-width: 640px)');
@@ -88,109 +196,5 @@ export const UpdateUser = ({ user, tenantId }: Props) => {
         </div>
       </DrawerContent>
     </Drawer>
-  );
-};
-
-const Form = ({ user, tenantId }: Props) => {
-  const [disabled, setDisabled] = useState<boolean>(true);
-
-  const {
-    register,
-    handleSubmit,
-    control,
-    formState: { isSubmitting, isDirty, isValid, errors },
-  } = useForm<Schema>({
-    resolver: zodResolver(updateUserClientSchema),
-    mode: 'onBlur',
-    defaultValues: {
-      name: user.name,
-      email: user.email,
-      role: user.role,
-    },
-  });
-  const { mutate } = useUpdateUser(user.id, tenantId);
-
-  const onSubmit: SubmitHandler<Schema> = (values) => {
-    mutate(values);
-  };
-
-  useEffect(() => {
-    setDisabled(isSubmitting || !isDirty || !isValid);
-  }, [isDirty, isSubmitting, isValid]);
-
-  return (
-    <form
-      onSubmit={handleSubmit(onSubmit)}
-      className="flex flex-col items-center gap-4"
-    >
-      <fieldset className="mx-auto flex w-11/12 flex-col gap-2">
-        <div className="flex flex-col gap-1">
-          <Field label={'Name'} value={'name'} error={errors.name?.message}>
-            <Input
-              {...register('name')}
-              withIcon
-              defaultValue={user.name}
-              icon={<UserIcon className="h-5 w-5" />}
-              type="name"
-              id="name"
-              placeholder="Name"
-            />
-          </Field>
-        </div>
-        <div className="flex flex-col gap-1">
-          <Field label={'Email'} value={'email'} error={errors.email?.message}>
-            <Input
-              {...register('email', { required: true })}
-              withIcon
-              defaultValue={user.email}
-              icon={<AtSign className="h-5 w-5" />}
-              type="email"
-              id="email"
-              placeholder="Email"
-            />
-          </Field>
-        </div>
-        {user.role !== 'tenant' && (
-          <div className="flex flex-col gap-1">
-            <Label
-              htmlFor="role"
-              className="lowercase"
-              style={{ fontVariant: 'small-caps' }}
-            >
-              Role
-            </Label>
-            <Controller
-              control={control}
-              name="role"
-              rules={{ required: true }}
-              render={({ field: { onChange } }) => (
-                <Select onValueChange={onChange} defaultValue={user.role}>
-                  <SelectTrigger id="role">
-                    <SelectValue placeholder="Select a role" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="user" className="pl-8">
-                      User
-                    </SelectItem>
-                    <SelectItem value="admin" className="pl-8">
-                      Admin
-                    </SelectItem>
-                  </SelectContent>
-                </Select>
-              )}
-            />
-          </div>
-        )}
-      </fieldset>
-      <Button
-        variant={disabled ? 'disabled' : 'primary'}
-        weight="semibold"
-        className="lowercase"
-        style={{ fontVariant: 'small-caps' }}
-        disabled={disabled}
-      >
-        Update
-      </Button>
-    </form>
   );
 };
