@@ -1,13 +1,15 @@
-import { Tenant, User } from '@prisma/client';
+import type { User } from '@prisma/client';
 
-export type UserWithTenant = User & {
-  tenant: Tenant;
+export type Me = User & {
+  tenant: {
+    logo: string;
+    company: string;
+  };
 };
 
 export type RegisterInfo = {
   company: string;
   companyEmail: string;
-  domain?: string | null;
   email: string;
   customerId?: string;
   plan: 'free' | 'startup' | 'enterprise';
