@@ -1,4 +1,6 @@
-import { Stripe, loadStripe } from '@stripe/stripe-js';
+import { loadStripe } from '@stripe/stripe-js';
+
+import type { Stripe } from '@stripe/stripe-js';
 
 let stripePromise: Promise<Stripe | null>;
 export const getStripe = () => {
